+++ src/modules/classes/components/Header.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <div className="mb-8">
+      <h1 className="text-3xl font-bold text-[#2C4A52] mb-2">My Classes</h1>
+      <p className="text-[#6B7280]">View your booked classes and track your progress</p>
+    </div>
+  )
+}
+
+export default Header
