+++ src/modules/classes/templates/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Classes from '../components'
+
+const ClassesTemplates = () => {
+  return (
+    <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
+      <Classes />
+    </div>
+  )
+}
+
+export default ClassesTemplates
