+++ src/modules/dashboard/components/WelcomeSection.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+const WelcomeSection = () => {
+  return (
+    <div>
+      <h2 className="text-3xl font-bold text-neutral-dark mb-2">
+        Good morning, Mary 👋
+      </h2>
+      <p className="text-neutral-medium">
+        Ready to start your health journey today?
+      </p>
+    </div>
+  )
+}
+
+export default WelcomeSection
