+++ src/modules/game/components/Header.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <div className="mb-8">
+      <h1 className="text-3xl font-bold text-neutral-dark mb-2">Health Games</h1>
+      <p className="text-neutral-medium">Make your health journey fun with interactive games</p>
+    </div>
+  )
+}
+
+export default Header
