+++ src/modules/game/templates/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Game from '../components'
+
+const GameTemplates = () => {
+  return (
+    <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
+      <Game />
+    </div>
+  )
+}
+
+export default GameTemplates
