+++ src/modules/help/components/Header.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <div className="mb-8">
+      <h1 className="text-3xl font-bold text-neutral-dark mb-2">Help & Support</h1>
+      <p className="text-neutral-medium">Get help and support for your health journey</p>
+    </div>
+  )
+}
+
+export default Header
