+++ src/modules/help/templates/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Help from '../components'
+
+const HelpTemplates = () => {
+  return (
+    <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
+      <Help />
+    </div>
+  )
+}
+
+export default HelpTemplates
