+++ src/modules/login/templates/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import Login from '../components'
+
+const LoginTemplates = () => {
+  return <Login />
+}
+
+export default LoginTemplates
