+++ src/modules/progress/components/Header.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Header = () => {
+  return (
+    <div className="mb-8">
+      <h1 className="text-3xl font-bold text-neutral-dark mb-2">Your Progress</h1>
+      <p className="text-neutral-medium">Track your health journey and celebrate achievements</p>
+    </div>
+  )
+}
+
+export default Header
