+++ src/modules/progress/templates/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Progress from '../components'
+
+const ProgressTemplates = () => {
+  return (
+    <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
+      <Progress />
+    </div>
+  )
+}
+
+export default ProgressTemplates
