+++ src/modules/register/templates/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import Register from '../components'
+
+const RegisterTemplates = () => {
+  return <Register />
+}
+
+export default RegisterTemplates
